Fix bug with raw fingerprint data
Integer values printed by fpcalc are 32-bit integers but were being
written to a buffer as 8-bit integers and thus losing some of the
fingerprint data.

This changes the buffer to have 32-bit signed big-endian integers to be
written to the buffer.

The original exposed string of integers is now also exposed as
`fingerprintRaw`

See https://github.com/parshap/node-fpcalc/pull/5#issuecomment-211910429